requirements.txt
keras
pandas
requests
tqdm
numpy
rasterio
geopandas
pillow
torchvision
torch
matplotlib
torchbearer
sklearn
texttable
black
pylint
pydocstyle
imblearn
xgboost
azure-storage-blob
